Introduce IAppBarItem (and InternalAppBarContext) to make it easier to add 'applications into the AppBar
