+++ examples/Demo/Shared/Microsoft.FluentUI.AspNetCore.Components.xml
@@ -319,63 +319,63 @@
         <member name="M:Microsoft.FluentUI.AspNetCore.Components.FluentAppBar.OverflowRaisedAsync(System.String)">
             <summary />
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.Href">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.Href">
             <summary>
             Gets or sets the URL for this item.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.Match">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.Match">
             <summary>
             Gets or sets how the link should be matched.
             Defaults to <see cref="F:Microsoft.AspNetCore.Components.Routing.NavLinkMatch.Prefix"/>.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.IconRest">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.IconRest">
             <summary>
             Gets or sets the Icon to use when the item is not hovered/selected/active.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.IconActive">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.IconActive">
             <summary>
             Gets or sets the Icon to use when the item is hovered/selected/active.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.Text">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.Text">
             <summary>
             Gets or sets the text to show under the icon.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.Tooltip">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.Tooltip">
             <summary>
             Gets or sets the tooltip to show when the item is hovered.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.Count">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.Count">
             <summary>
             Gets or sets the count to show on the item with a <see cref="T:Microsoft.FluentUI.AspNetCore.Components.FluentCounterBadge"/>.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.Order">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.Order">
             <summary>
             Gets or sets the position of the item in the app bar.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.ChildContent">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.ChildContent">
             <summary>
              Gets or sets the content to be shown.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.OnClick">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.OnClick">
             <summary>
             The callback to invoke when the item is clicked.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.Owner">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.Owner">
             <summary>
             Gets or sets the owning FluentAppBar component.
             </summary>
         </member>
-        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItemBase.Overflow">
+        <member name="P:Microsoft.FluentUI.AspNetCore.Components.FluentAppBarItem.Overflow">
             <summary>
             If this app is outside of visible app bar area.
             </summary>
