+++ src/Core/Components/AppBar/FluentAppBarItem.razor
@@ -1,5 +1,5 @@
 ﻿@namespace Microsoft.FluentUI.AspNetCore.Components
-@inherits FluentAppBarItemBase
+@inherits FluentComponentBase
 
 <div id=@Id @attributes="AdditionalAttributes" class="@ClassValue" style="@Style" title="@(Tooltip ?? Text)">
     <FluentCounterBadge Count="@Count"
